// apps/trade-pro/src/pages/SaleTrading/saleOrder/type.ts
import exp from 'constants';

export type TSaleOrderDetail = {
  Id: number;
  OrderItemId: number;
  OrderItemUOMId: number;
  OrderItemQty: number;
  NetWeight: number;
  RateAddLess: number;
  RateWithouAdLs: number;
  OrderItemRate: number;
  OrderItemRateUOMId: number;
  Amount: number;
  OrderRemarks: number;
  ActionTypeId: number;
  DocumentTypeId: number;
  DocDate: string;
  OrganizationId: number;
  CompanyId: number;
  BranchesId: number;
  FinancialYearId: number;
  EntryUser: number;
  EntryDate: string;
  ModifyUser: number;
  ModifyDate: string;
  DeliveryStartDate: string;
  DocNo: number;
  OrderSupCustId: number;
  RemarksHeader: string;
  PaymentTermsId: number;
  OrderDueDays: number;
  OrderDueDate: string;
  OrderExpiryDate: string;
  DeliveryTerm: string;
  ShiptoAddress: number;
  SubPartyId: number;
  BrokerAgentSupCustId: number;
  CommissionType: string;
  CommRate: number;
  CommAmount: number;
  CommissionRemarks: string;
  SaleOrderDetailList: TSaleOrderDetail[];
};

export type TPurchaseOrderDetailEntry = {
  Id?: number;
  LineId?: number;
  UOMCode?: string;
  RateUom?: string;
  BagPrice?: number;
  ItemName?: string;
  BagWeight?: number;
  TaxAmount?: number;
  TaxNameId?: number;
  RetailRate?: number;
  TaxPercent?: number;
  TotalAmount?: number;
  Crop?: string | null;
  ActionTypeId?: number;
  Amount?: number | null;
  OrderItemRate?: number;
  OrderItemUOMId?: number;
  TaxName?: string | null;
  UOMDescription?: number;
  JobLotId?: null | number;
  PurchaseOrderId?: number;
  CityArea?: string | null;
  Moisture?: string | null;
  NetWeight?: number | null;
  JobLotDescription?: string;
  OrderItemId?: null | number;
  OrderItemRateUOMId?: number;
  LabSampleNo?: number | null;
  OrderItemQty?: number | null;
  OrderRemarks?: string | null;
  TaxableStatus?: string | null;
  EquivalentRate?: number | null;
  AmountCalcType?: string | null;
  InvLabSampleAnalysisHeaderId?: number;
};

export type TDetailItem = {
  Id: number;
  UOMCode: string;
  ItemName: string;
  ItemUomId: number;
  Equivalent: number;
};

export type TPaymentTerms = {
  Id: number;
  CompanyId: number;
  TermsCode: string;
  PostDate: Date | null;
  EntryDate: Date | null;
  OrganizationId: number;
  ModifyDate: Date | null;
  TermsDescription: string;
  PostState: string | null;
  PostUser: number | string | null;
  EntryUser: number | string | null;
  ModifyUser: number | string | null;
};

export type TSaleOrder = {
  Edit: string;
  Print: string;
  SaveAs: string;
  DocNo: number;
  DocDate: Date;
  Customer: string;
  SubParty_Account: '';
  ShipToAddress: '';
  DueDate: Date;
  DeliveryTerm: string;
  PaymentTerm: string;
  CommissionAgent: '';
  CommissionType: '';
  CommissionRate: number;
  CommissionAmount: number;
  CommissionRemarks: '';
  status: '';
  EntryDate: Date;
  IsApproved: string;
  EntryUser: string;
  Remarks: '';
};
export type TSaleOrderForm = {
  ItemName: string;
  PackUom: string | number;
  QTY: number;
  Weight: number;
  Rate: number;
  RateUOM: number | string;
  Amount: number;
  Remarks: '';
};
export type TSaleOrderForm2 = {};